requirements.txt
aniso8601==4.0.1
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
Click==7.0
coverage==4.5.2
Flask==1.0.2
Flask-API==1.0
Flask-RESTful==0.3.7
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
pkg-resources==0.0.0
pluggy==0.8.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
py==1.7.0
PyJWT==1.7.1
pylint==2.2.2
pytest==4.0.2
pytz==2018.7
six==1.12.0
typed-ast==1.1.1
Werkzeug==0.14.1
wrapt==1.11.0
